Update patch set 1
Patch Set 1:

The same change is already present in froyo and gingerbread branches, but not in ics and jellybean: https://github.com/CyanogenMod/android_packages_apps_SoundRecorder/commit/935b5e25d3c28d634922c1a66a95c8d275ace55a (but the commit message there looks better). So the problem solved by this patch may even be considered a regression in ics and jellybean branches.

However, is android.intent.category.DEFAULT really needed in this intent filter? It is documented to be used for activities which perform actions on some data, which does not apply to starting an app from the launcher; adding just android.intent.category.LAUNCHER should be enough to make Sound Recorder appear in the launcher.

Patch-set: 1