Update patch set 1
Patch Set 1: Code-Review+1

the thing is how to notify the called person assuming that not everyone is on cm :p

at least the only solution I have in mind currently is to use tts or pico engine and give the other party a hint that recording has started. but still that does not cover the countries where the law explicit says that the called person needs to agree on it.

other then that yes feature wise it is of course really nice

Patch-set: 1