Update patch set 1
Patch Set 1:

Don't forget there's a lot of countries that have no law restrictions on phone call recording.

If we're making this an option -- its purpose is to hide that "Record" button by default. It doesn't depend on whether the other party is on CM or not -- we shouldn't really care about this. If the user enables the Call Recording option and agrees with our usual warning -- well, he is informed, it's his responsibility from now on. We assume he will inform the other party by himself.

We absolutely should _NOT_ make any TTS notifications or even try to imagine how to get a confirmation from the other party. Again, there's a lot of countries where phone call recording is permitted.

Patch-set: 1