Update patch set 1
Patch Set 1:

at least I am not a lawyer and cannot say if it is enough in this countries to move the responsability to the user tbh. So maybe it would be good a lawyer can comment on this thing.

Of course if it is enough then a popup hint combined with an option is the way to go.

I am just carfull when it comes to law issues....especially looking on the fact that cm is a company now.....here in germany eg they are really picky on such things lol

Patch-set: 1