Update patch set 2
Patch Set 2:

Maybe I am wrong, but if you have a line of code which is using a string resource, you tell it to look for the string in "@string" and for CM strings in "@cm_strings" right? If I put all the translations in "cm_strings" would the app know where to find it?

Patch-set: 2