Update patch set 4
Patch Set 4: Code-Review-1

The problem is in the intent - the Extras I put in are specific to CMFileManager so it knows what folder to open. Opening this in any other app would do nothing since it doesn't know how to handle those extras. I think the proper way to do this would be to get the proper URI for the location of the recordings and put that in the intent data.

Patch-set: 4