Update patch set 4
Patch Set 4:

Maybe something like this: http://pastebin.com/i30GHecB

But then there's still a few problems: it opens the document UI, but doesn't open it to the proper folder. I don't think that URI is what the documents UI is expecting so it doesn't show the proper folder, I'm not sure off the top of my head how to get that URI.

Alternatively if we don't go the documents UI way, the problem is you need to find a common intent action that will work for every file browser, but I'm not sure if there is one, it needs some investigating.

What file manager are you trying to use this with?

Patch-set: 4