Update patch set 11
Patch Set 11:

Hmm.. as far as my understanding goes, this option is not to pick an audio file and open it in the app... It just to open the folder where the recording saves to, and then you can do whatever you want with them (Share, play in a player).... there is no point of opening the recording again in this app, as it will do nothing. This app is not an editor/audio player. I just want to open the output folder in a file manager of the user's pick. That's all. What do you think is the best way to do that?

Patch-set: 11