Update patch set 16
Patch Set 16: Code-Review+1

Finally! This patch does it all: if there is only one file manager it will open the directory with it, and if more than one, it will ask to pick, and only between file managers. Checked and confirmed working great with ES explorer and CM file manager.

Patch-set: 16