Update patch set 18
Patch Set 18:

Hmm... Fixing this in the CM manager is not something I know how to code at the moment... We'll need much more experienced java developer for that. This is a good fix for now, and when the support will be added I'll remove the second part. But for now, it's much better than what used to be.

Patch-set: 18