Update patch set 6
Patch Set 6:

(12 comments)

I like the new UI, well done.

However, I miss the visual rendering of the microphone level,
especially given the fact it's a way of testing the microphone works.

I would imagine a small animation on the levels drawable beside the "Recording" test, lowering and increasing with the volume.

But that's just pure cosmetic and visual confirmation of course.

Patch-set: 6