Update patch set 4
Patch Set 4:

May be the user wants to send an email with the old recorded audio file form the file manager during audio recording. Or copy it to another folder.  The 'View recordings' just goto the file manager to show the files, not playback it immediately. If user wants to playback that audio file manually, it is user's choice. User can also playback music while doing audio recording also. There is nothing wrong with this patch, more than a matter of taste.

Patch-set: 4